requirements.txt
aiohttp==3.9.1
python-dateutil==2.8.2
